+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Improve `stud items:show` command to display description in dedicated sections with automatic text sanitization and divider detection [SCI-8]
+
 ## [2.4.1] - 2025-11-25
 
 ### Removed